# api/openapi.yaml
openapi: 3.0.3
info:
  title: Micro-Blog API
  version: 0.1.0
  description: |
    Authenticated markdown micro-blog with a built-in URL shortener.
    Dual-token JWT (15min access + 7d refresh-cookie); RBAC enforced by a
    centralized permission/scope matrix.
servers:
  - url: http://localhost:8090
    description: Local dev (PORT from .env)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/refresh.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly refresh token cookie issued by /auth/login.
  parameters:
    PathID:
      name: id
      in: path
      required: true
      description: Numeric primary key.
      schema:
        type: integer
        format: int64
        minimum: 1
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid_input
          description: Machine-readable error code.
        message:
          type: string
          example: validation failed
          description: Human-readable detail (optional).
        fields:
          type: object
          additionalProperties:
            type: string
          example:
            username: must be at least 3 characters
            email: is not a valid email
          description: |
            Per-field validation messages, present only on `invalid_input`
            responses where one or more fields failed validation. Each value
            is a single human-readable message for that field.
    User:
      type: object
      required: [id, username, email, role_id, role]
      properties:
        id: { type: integer, format: int64, example: 1 }
        username: { type: string, example: alice }
        email: { type: string, format: email, example: alice@example.com }
        role_id: { type: integer, format: int64, example: 3 }
        role: { type: string, example: Author }
...
    UserUpdateRequest:
      type: object
      description: Partial update; any omitted field stays unchanged.
      properties:
        username: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        role_id:
          type: integer
          format: int64
          enum: [1, 2, 3, 4]
...
  responses:
    BadRequest:
      description: Bad request — invalid body or input.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Authentication required or token invalid/expired.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Uniqueness constraint violation (username/email).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /admin/users/{id}:
    get:
      tags: [admin-users]
      operationId: getUser
      x-roles: [Admin]
      summary: Read a user. Admin only.
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '200':
          description: The user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [admin-users]
      operationId: updateUser
      x-roles: [Admin]
      summary: Partial update of any field (incl. role_id). Admin only.
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200':
          description: Updated user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [admin-users]
      operationId: deleteUser
      x-roles: [Admin]
      summary: Delete a user. Admin only; cannot delete self.
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '204': { description: Deleted. }
        '400':
          description: Refused — Admins cannot delete themselves.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }